requirements.txt
hopsworks==3.7.0
pandas
scikit-learn
streamlit
requests
python-dotenv
joblib